requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
boto3==1.24.57
botocore==1.27.57
certifi==2022.6.15.1
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
crispy-tailwind==0.5.0
cryptography==38.0.1
decorator==5.1.1
defusedxml==0.7.1
Django==4.1
django-allauth==0.51.0
django-captcha==1.0.0
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-environ==0.9.0
django-filter==22.1
django-recaptcha==3.0.0
django-social-share==2.3.0
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
environ==1.0
frozenlist==1.3.1
future==0.18.2
GDAL==3.5.3
geocoder==1.38.1
geographiclib==1.52
geoip2==4.6.0
geopy==2.2.0
idna==3.3
jmespath==1.0.1
maxminddb==2.2.0
multidict==6.0.2
oauthlib==3.2.1
Pillow==9.2.0
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.4.0
PyMySQL==1.0.2
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.2.1
q==2.7
qrcode==7.3.1
ratelim==0.1.6
requests==2.28.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.12
yarl==1.8.1
